changed doubles to floats

+++ src/Math.cxx
@@ -23,18 +23,18 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301 USA
 
 // fft routines adapted from those found here:
 // http://www.dspguide.com/ch12/3.htm
-void Math::forwardFFT(double *real, double *imag, int size)
+void Math::forwardFFT(float *real, float *imag, int size)
 {
   int j = size / 2;
   int i = 0, k = 0;
-  const double pi = 3.14159265;
+  const float pi = 3.14159265;
 
   for(i = 1; i <= size - 2; i++)
   {
     if(i < j)
     {
-      const double tr = real[j];
-      const double ti = imag[j];
+      const float tr = real[j];
+      const float ti = imag[j];
       real[j] = real[i];
       imag[j] = imag[i];
       real[i] = tr;
@@ -52,37 +52,37 @@ void Math::forwardFFT(double *real, double *imag, int size)
     j += k;
   }
 
-  for(k = 1; k <= (int)(log(size) / log(2)); k++)
+  for(k = 1; k <= (int)(logf(size) / logf(2)); k++)
   {
-    const int le = (int)(pow(2, k));
+    const int le = (int)(powf(2, k));
     const int le2 = le / 2;
 
-    double ur = 1;
-    double ui = 0;
-    double sr = cos(pi / le2);
-    double si = -sin(pi / le2);
+    float ur = 1;
+    float ui = 0;
+    float sr = cosf(pi / le2);
+    float si = -sinf(pi / le2);
 
     for(j = 1; j <= le2; j++)
     {
       for(i = j - 1; i < size; i += le)
       {
         const int ip = i + le2;
-        const double tr = real[ip] * ur - imag[ip] * ui;
-        const double ti = real[ip] * ui + imag[ip] * ur;
+        const float tr = real[ip] * ur - imag[ip] * ui;
+        const float ti = real[ip] * ui + imag[ip] * ur;
         real[ip] = real[i] - tr;
         imag[ip] = imag[i] - ti;
         real[i] += tr;
         imag[i] += ti;
       }
 
-      const double tr = ur;
+      const float tr = ur;
       ur = tr * sr - ui * si;
       ui = tr * si + ui * sr;
     }
   }
 }
 
-void Math::inverseFFT(double *real, double *imag, int size)
+void Math::inverseFFT(float *real, float *imag, int size)
 {
   for(int k = 0; k < size; k++)
     imag[k] = -imag[k];
